incorporate dennis comment
(1) instead of just printing out the number of email addresses in the
original input file and the number of accounts found, please print out
all the email addresses that could not be matched.
(2) for the ones that do match, save the objects into a separate array
so that if “.y” is entered, the array of user account JSON objects can
be written to disk (a kind of backup log in case something goes wrong).
(3) instead of passing the emailAddress’s into deleteMany(), i’d like
you to use the _id’s from the results of the initial find() call. that
is a lot safer and protects against the theoretical possibility that one
of the email addresses doesn’t originally match but before you press .y,
a new user registers with that email address and then is subsequently
deleted even though the administrator was never told that the account
would be deleted.